use filterblur as parent class
